Added whitelist feature, added ability to change items for shop on the fly and refactored code

+++ org/wargamer2010/essentials/SetExpFix.java
@@ -73,7 +73,7 @@ public class SetExpFix
 		}
 		return 17;
 	}
-	
+
 	public static int getExpToLevel(final int level)
 	{
 		int currentLevel = 0;
@@ -91,7 +91,7 @@ public class SetExpFix
 	//Without this people would be able to use exp and then still sell it.
 	public static int getTotalExperience(final Player player)
 	{
-		int exp = (int)Math.round(getExpAtLevel(player) * player.getExp());
+		int exp = Math.round(getExpAtLevel(player) * player.getExp());
 		int currentLevel = player.getLevel();
 
 		while (currentLevel > 0)
@@ -101,10 +101,10 @@ public class SetExpFix
 		}
 		return exp;
 	}
-	
+
 	public static int getExpUntilNextLevel(final Player player)
 	{
-		int exp = (int)Math.round(getExpAtLevel(player) * player.getExp());		
+		int exp = Math.round(getExpAtLevel(player) * player.getExp());
 		int nextLevel = player.getLevel();
 		return getExpAtLevel(nextLevel) - exp;
 	}
